LayoutTests/imported/w3c/web-platform-tests/wasm/jsapi/interface.any.worker-expected.txt

CONSOLE MESSAGE: line 154: TypeError: undefined is not an object (evaluating 'WebAssembly.Global.prototype')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'WebAssembly.Global.prototype')

PASS WebAssembly: property descriptor 
PASS WebAssembly: calling 
PASS WebAssembly: constructing 
PASS WebAssembly.Module: property descriptor 
PASS WebAssembly.Module: prototype 
PASS WebAssembly.Module: prototype.constructor 
PASS WebAssembly.Instance: property descriptor 
PASS WebAssembly.Instance: prototype 
PASS WebAssembly.Instance: prototype.constructor 
PASS WebAssembly.Memory: property descriptor 
PASS WebAssembly.Memory: prototype 
PASS WebAssembly.Memory: prototype.constructor 
PASS WebAssembly.Table: property descriptor 
PASS WebAssembly.Table: prototype 
PASS WebAssembly.Table: prototype.constructor 
FAIL WebAssembly.Global: property descriptor assert_equals: expected "object" but got "undefined"
FAIL WebAssembly.Global: prototype undefined is not an object (evaluating 'Object.getOwnPropertyDescriptor(interface_object, "prototype")')
FAIL WebAssembly.Global: prototype.constructor undefined is not an object (evaluating 'interface_object.prototype')
PASS WebAssembly.CompileError: property descriptor 
PASS WebAssembly.CompileError: prototype 
PASS WebAssembly.CompileError: prototype.constructor 
PASS WebAssembly.LinkError: property descriptor 
PASS WebAssembly.LinkError: prototype 
PASS WebAssembly.LinkError: prototype.constructor 
PASS WebAssembly.RuntimeError: property descriptor 
PASS WebAssembly.RuntimeError: prototype 
PASS WebAssembly.RuntimeError: prototype.constructor 
PASS WebAssembly.validate 
PASS WebAssembly.validate: name 
PASS WebAssembly.validate: length 
PASS WebAssembly.compile 
PASS WebAssembly.compile: name 
PASS WebAssembly.compile: length 
PASS WebAssembly.instantiate 
PASS WebAssembly.instantiate: name 
PASS WebAssembly.instantiate: length 
FAIL WebAssembly.Module.exports assert_equals: expected (undefined) undefined but got (function) function "function exports() {
    [native code]
}"
PASS WebAssembly.Module.exports: name 
PASS WebAssembly.Module.exports: length 
FAIL WebAssembly.Module.imports assert_equals: expected (undefined) undefined but got (function) function "function imports() {
    [native code]
}"
PASS WebAssembly.Module.imports: name 
PASS WebAssembly.Module.imports: length 
FAIL WebAssembly.Module.customSections assert_equals: expected (undefined) undefined but got (function) function "function customSections() {
    [native code]
}"
PASS WebAssembly.Module.customSections: name 
PASS WebAssembly.Module.customSections: length 
PASS WebAssembly.Instance.exports 
PASS WebAssembly.Instance.exports: getter 
PASS WebAssembly.Instance.exports: setter 
FAIL WebAssembly.Memory.grow assert_equals: expected (undefined) undefined but got (function) function "function grow() {
    [native code]
}"
PASS WebAssembly.Memory.grow: name 
PASS WebAssembly.Memory.grow: length 
PASS WebAssembly.Memory.buffer 
PASS WebAssembly.Memory.buffer: getter 
PASS WebAssembly.Memory.buffer: setter 
FAIL WebAssembly.Table.grow assert_equals: expected (undefined) undefined but got (function) function "function grow() {
    [native code]
}"
PASS WebAssembly.Table.grow: name 
PASS WebAssembly.Table.grow: length 
FAIL WebAssembly.Table.get assert_equals: expected (undefined) undefined but got (function) function "function get() {
    [native code]
}"
PASS WebAssembly.Table.get: name 
PASS WebAssembly.Table.get: length 
FAIL WebAssembly.Table.set assert_equals: expected (undefined) undefined but got (function) function "function set() {
    [native code]
}"
PASS WebAssembly.Table.set: name 
PASS WebAssembly.Table.set: length 
PASS WebAssembly.Table.length 
PASS WebAssembly.Table.length: getter 
PASS WebAssembly.Table.length: setter 
